+++ src/validators/enrollment.rs
@@ -35,7 +35,11 @@ pub fn validate_enrollment(
         Err(e) => {
             return Err(Box::new(Error::new(
                 ErrorKind::Other,
-                format!("Json parsing failed with error {}", e.to_string()),
+                format!(
+                    "{} parsing failed with error {}",
+                    config_file_name.to_string_lossy(),
+                    e.to_string()
+                ),
             )));
         }
         Ok(body) => body,
@@ -52,4 +56,60 @@ pub fn validate_enrollment(
         }
     }
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::PathBuf;
+
+    #[test]
+    fn enrollment_config_missing_dps() {
+        assert_ne!(
+            None,
+            validate_enrollment(&PathBuf::from(
+                "testfiles/enrollment_static_missing_dps.json"
+            ),)
+            .unwrap_err()
+            .to_string()
+            .find("missing field `dpsConnectionString`")
+        );
+    }
+
+    #[test]
+    fn enrollment_config_missing_iothub() {
+        assert_ne!(
+            None,
+            validate_enrollment(&PathBuf::from(
+                "testfiles/enrollment_static_missing_iothub.json"
+            ),)
+            .unwrap_err()
+            .to_string()
+            .find("missing field `iothubConnectionString`")
+        );
+    }
+
+    #[test]
+    fn enrollment_config_unknown_key() {
+        assert_eq!(
+            true,
+            validate_enrollment(&PathBuf::from(
+                "testfiles/enrollment_static_unknown_key.json"
+            ),)
+            .is_ok()
+        );
+    }
+
+    #[test]
+    fn enrollment_config_invalid_connection_string() {
+        assert_ne!(
+            None,
+            validate_enrollment(&PathBuf::from(
+                "testfiles/enrollment_static_invalid_connection_string.json"
+            ),)
+            .unwrap_err()
+            .to_string()
+            .find("Enrollment validation failed")
+        );
+    }
 }